[GN-126] refactor: SortDropdown 파일 경로 변경
