+++ src/app/(DefaultLayout)/(MainLayout)/page.tsx
@@ -6,9 +6,9 @@ import { useEffect, useState } from "react";
 import ArrowButton from "./components/ArrowButton";
 import CategoryFilter from "./components/CategoryFilter";
 import activityService from "@/apis/activity/activity.service";
+import SortDropdown from "@/app/(DefaultLayout)/(MainLayout)/components/SortDropdown";
+import type { SortOptionType } from "@/app/(DefaultLayout)/(MainLayout)/components/SortDropdown";
 import Card from "@/components/card/Card";
-import SortDropdown from "@/components/dropdown/SortDropdown";
-import type { SortOptionType } from "@/components/dropdown/SortDropdown";
 import Pagination from "@/components/pagination/Pagination";
 import { Search } from "@/components/search/Search";
 import type { ActivityCategoryType } from "@/types/activity";
